Safely manage creating files in scripts.
Use a temporary directory with a unique name to read/write files into
during the duration of the script. Once the script is done the temporary
directory is automatically cleaned up.

If, however, the script is unsuccessful (i.e., signal is not EXIT (0))
then the directory will be in the /tmp directory where you can find all
the files that you created/modified/downloaded.